CharacterStamp is now distinct from SignatureStamp, which can be used by a DataSource.
